+++ src/main/java/ru/dpopkov/tasktracker/model/BaseEntity.java
@@ -1,0 +1,9 @@
+package ru.dpopkov.tasktracker.model;
+
+import lombok.Data;
+
+@Data
+public abstract class BaseEntity {
+
+    private Long id;
+}