+++ src/main/java/ru/dpopkov/tasktracker/model/Project.java
@@ -1,16 +1,19 @@
 package ru.dpopkov.tasktracker.model;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 import java.util.HashSet;
 import java.util.Set;
 
 @Data
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @NoArgsConstructor
-public class Project {
+public class Project extends BaseEntity {
 
-    private Long id;
     private String name;
     private String description;
     private Set<User> users = new HashSet<>();
@@ -30,3 +33,4 @@ public class Project {
         user.removeProject(this);
     }
 }
+