+++ src/main/java/ru/dpopkov/tasktracker/repositories/ProjectRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.repositories;
+
+import ru.dpopkov.tasktracker.model.Project;
+
+public interface ProjectRepository extends BaseRepository<Project> {
+}