+++ src/main/java/ru/dpopkov/tasktracker/repositories/TaskRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.repositories;
+
+import ru.dpopkov.tasktracker.model.Task;
+
+public interface TaskRepository extends BaseRepository<Task> {
+}