+++ src/main/java/ru/dpopkov/tasktracker/repositories/UserRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.repositories;
+
+import ru.dpopkov.tasktracker.model.User;
+
+public interface UserRepository extends BaseRepository<User> {
+}