+++ src/main/java/ru/dpopkov/tasktracker/repositories/inmemory/SimpleProjectRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.repositories.inmemory;
+
+import ru.dpopkov.tasktracker.model.Project;
+
+public class SimpleProjectRepository extends SimpleBaseRepository<Project> {
+}