+++ src/main/java/ru/dpopkov/tasktracker/repositories/inmemory/SimpleTaskRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.repositories.inmemory;
+
+import ru.dpopkov.tasktracker.model.Task;
+
+public class SimpleTaskRepository extends SimpleBaseRepository<Task> {
+}