+++ src/main/java/ru/dpopkov/tasktracker/repositories/inmemory/SimpleUserRepository.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.repositories.inmemory;
+
+import ru.dpopkov.tasktracker.model.User;
+
+public class SimpleUserRepository extends SimpleBaseRepository<User> {
+}