+++ History.md
@@ -4,3 +4,4 @@ History of implementation
 1. Add empty starting project.
 2. Add config for unit and integration tests.
 3. Add Model: User, Project, Task.
+4. Add interfaces for repositories and simple in-memory implementations.